# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quiz app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface QuizSummary {
  id: number;
  title: string;
  description: string;
  category: string;
  questionCount: number;
  attemptCount: number;
  createdAt: string;
}

export interface Choice {
  id: number;
  text: string;
}

export interface Question {
  id: number;
  quizId: number;
  text: string;
  choices: Choice[];
  correctChoiceId: number;
  order: number;
}

export interface Quiz {
  id: number;
  title: string;
  description: string;
  category: string;
  questions: Question[];
  createdAt: string;
}

export interface CreateQuizBody {
  title: string;
  description: string;
  category: string;
}

export interface UpdateQuizBody {
  title?: string;
  description?: string;
  category?: string;
}

export interface CreateQuestionBody {
  text: string;
  /** @minItems 2 */
  choices: string[];
  correctChoiceIndex: number;
}

export interface Answer {
  questionId: number;
  choiceId: number;
}

export interface SubmitAttemptBody {
  playerName: string;
  answers: Answer[];
}

export interface AnswerResult {
  questionId: number;
  questionText: string;
  choiceId: number;
  choiceText: string;
  correctChoiceId: number;
  correctChoiceText: string;
  isCorrect: boolean;
}

export interface AttemptResult {
  id: number;
  quizId: number;
  playerName: string;
  score: number;
  total: number;
  percentage: number;
  answers: AnswerResult[];
  completedAt: string;
}

export interface QuizStats {
  quizId: number;
  totalAttempts: number;
  averageScore: number;
  topScore: number;
  topScorer: string;
}

export interface RecentAttempt {
  playerName: string;
  quizTitle: string;
  score: number;
  total: number;
  percentage: number;
  completedAt: string;
}

export interface SummaryStats {
  totalQuizzes: number;
  totalAttempts: number;
  totalPlayers: number;
  recentAttempts: RecentAttempt[];
}