# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quiz app API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all quizzes
 */
export const ListQuizzesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  questionCount: zod.number(),
  attemptCount: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListQuizzesResponse = zod.array(ListQuizzesResponseItem);

/**
 * @summary Create a new quiz
 */
export const CreateQuizBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
});

/**
 * @summary Get a quiz with all questions and answers
 */
export const GetQuizParams = zod.object({
  id: zod.coerce.number(),
});

export const GetQuizResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  questions: zod.array(
    zod.object({
      id: zod.number(),
      quizId: zod.number(),
      text: zod.string(),
      choices: zod.array(
        zod.object({
          id: zod.number(),
          text: zod.string(),
        }),
      ),
      correctChoiceId: zod.number(),
      order: zod.number(),
    }),
  ),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a quiz
 */
export const UpdateQuizParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateQuizBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  category: zod.string().optional(),
});

export const UpdateQuizResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  questions: zod.array(
    zod.object({
      id: zod.number(),
      quizId: zod.number(),
      text: zod.string(),
      choices: zod.array(
        zod.object({
          id: zod.number(),
          text: zod.string(),
        }),
      ),
      correctChoiceId: zod.number(),
      order: zod.number(),
    }),
  ),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a quiz
 */
export const DeleteQuizParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Add a question to a quiz
 */
export const AddQuestionParams = zod.object({
  id: zod.coerce.number(),
});

export const addQuestionBodyChoicesMin = 2;

export const AddQuestionBody = zod.object({
  text: zod.string(),
  choices: zod.array(zod.string()).min(addQuestionBodyChoicesMin),
  correctChoiceIndex: zod.number(),
});

/**
 * @summary Update a question
 */
export const UpdateQuestionParams = zod.object({
  id: zod.coerce.number(),
});

export const updateQuestionBodyChoicesMin = 2;

export const UpdateQuestionBody = zod.object({
  text: zod.string(),
  choices: zod.array(zod.string()).min(updateQuestionBodyChoicesMin),
  correctChoiceIndex: zod.number(),
});

export const UpdateQuestionResponse = zod.object({
  id: zod.number(),
  quizId: zod.number(),
  text: zod.string(),
  choices: zod.array(
    zod.object({
      id: zod.number(),
      text: zod.string(),
    }),
  ),
  correctChoiceId: zod.number(),
  order: zod.number(),
});

/**
 * @summary Delete a question
 */
export const DeleteQuestionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Submit a quiz attempt with answers
 */
export const SubmitAttemptParams = zod.object({
  id: zod.coerce.number(),
});

export const SubmitAttemptBody = zod.object({
  playerName: zod.string(),
  answers: zod.array(
    zod.object({
      questionId: zod.number(),
      choiceId: zod.number(),
    }),
  ),
});

/**
 * @summary List recent attempts for a quiz
 */
export const ListAttemptsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListAttemptsResponseItem = zod.object({
  id: zod.number(),
  quizId: zod.number(),
  playerName: zod.string(),
  score: zod.number(),
  total: zod.number(),
  percentage: zod.number(),
  answers: zod.array(
    zod.object({
      questionId: zod.number(),
      questionText: zod.string(),
      choiceId: zod.number(),
      choiceText: zod.string(),
      correctChoiceId: zod.number(),
      correctChoiceText: zod.string(),
      isCorrect: zod.boolean(),
    }),
  ),
  completedAt: zod.coerce.date(),
});
export const ListAttemptsResponse = zod.array(ListAttemptsResponseItem);

/**
 * @summary Get stats for a quiz (attempts count, avg score, top score)
 */
export const GetQuizStatsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetQuizStatsResponse = zod.object({
  quizId: zod.number(),
  totalAttempts: zod.number(),
  averageScore: zod.number(),
  topScore: zod.number(),
  topScorer: zod.string(),
});

/**
 * @summary Overall app-wide stats (total quizzes, attempts, top players)
 */
export const GetSummaryStatsResponse = zod.object({
  totalQuizzes: zod.number(),
  totalAttempts: zod.number(),
  totalPlayers: zod.number(),
  recentAttempts: zod.array(
    zod.object({
      playerName: zod.string(),
      quizTitle: zod.string(),
      score: zod.number(),
      total: zod.number(),
      percentage: zod.number(),
      completedAt: zod.coerce.date(),
    }),
  ),
});